Remove POC of core plugins, replace with new plugin paradigm.
